New copy constructor (not documented) in Triangulation_ds_cell_3.h
Only used by Triangulation_data_structure_3.h in copy_tds to copy cells
while keeping the possible additional information in the cell base
but adding it to the list of the new tds.

Takes as arguments : the tds, 4 vertices, and the old cell.

Similar modification in copy_tds to copy vertices but the standard
copy constructor is enough to achieve the goal (no list of vertices...).