changed \begin{ccClassTemplate}{Circulator_from_iterator<I>} to \begin{ccRefClass}{Circulator_from_iterator<I>} following the message from Susan Hello, Mariette,
When I incorporated the modified Circulator_stl.tex file that Lutz sent
into the "Use of STL" manual, there was one small problem that I thought
you might like to fix.  It occurs around line 369 in the following command:

\begin{ccClassTemplate}{Circulator_from_iterator<I>}

The problem occurs because the <I> looks (at times) to HTML like the beginning
of an italicized section of text, with the result that everything following
this is in italics.  I noticed it first in the index for this manual, but
it may have occurred also in the actual text.  Anyway, I fixed it by changing
the command to:

\begin{ccRefClass}{Circulator_from_iterator<I>}

(where ccRefClass is one of the new manual commands).  Another option would
be to change the "I" to something like "It", but this is not as appealing
since it implies changes elsewhere in the text as well.