Boolean operations on polygons. We provide the following boolean operations on two polygons: 1. intersection. 2. union. 3. symmeteric diff 4. diff
the funcitons are templated with the polygon type and traits.
The CGAL polygon is a model of our polygon.
The traits is used for the sweep line algorithm.

Each function has three (optional) parameters representing the resulting features.
1. a list of points.
2. a list of segments.
3. a list of polygons.

The latter is a list of simple polygons (from the polygon type).
Hence if the we polygon with holes we split it to simple polygons.