* new feature: sub-manuals for the html-converter. own aux-files are generated, to reflect the difference in chapter-numbering between pdf/ps and html. aux-files currently only contain a list of counters and their values. Actually, the html-converter .aux files are called .hax (html-aux)
* the handling of \includeonly moved to latex.sty instead of lexer.lex. thus,
lciInput was no longer required to be a reserved macro -> removed

* renamed OpenFile into OpenTmpFile, because that's what it does.
a new function OpenFile was added, that allows  to open arbitrary files
(for example, .hax files for the html-converter).

* bugfix in latex.sty: now, each time a new part is started, the chapter-flag
is reset such that it looks like there was no chapter before. previously, there
was a bug leading to a stack-underflow upon \chapter{....}