Use install(PROGRAMS...) to install scripts, instead of install(DIRECTORY..) with the PERMISSIONS statement.
== Reason: that mail ==

Message-ID: <4A2D1852.5050308@itis.ethz.ch>
Date: Mon, 08 Jun 2009 15:55:30 +0200
From: Dominik Szczerba <dominik@itis.ethz.ch>
To: cgal-discuss@lists-sop.inria.fr
Subject: [cgal-discuss] small installation issue on linux (CGAL v3.4)
Reply-To: cgal-discuss@lists-sop.inria.fr

Hi,
While 'make install' a following error is issued:

CMake Error at cmake_install.cmake:55 (FILE):
   file Problem setting permissions on directory "/usr/local/bin"

The properties of this folder are:

drwxrwxr-x  2 root  staff 4096 2009-06-08 15:46 bin

which is certainly sufficient for any local install by a staff member. 
By temporarilly changing the ownership of the folder to myself the 
installation proceeds, but it changes the permissions to 'drwxr-xr-x', 
which is not how it should be done. Is there a particular reason to do 
it this way or is it simply overlooking?

with regards,
Dominik
-- 
You are currently subscribed to cgal-discuss.
To unsubscribe or access the archives, go to
https://lists-sop.inria.fr/wws/info/cgal-discuss