Add macro CGAL_MSVC_USE_STD_FABS
On Windows, with MSVC, std::fabs is an intrinsic function with /Oi (implied
with /O1, /O2, and so on) *if* /fp:strict is not used. We should reevaluate
the need to use /fp:strict (and maybe use /fp:precise, the default,
combined with #pragma fenv_access).

First step: bench! This macro will help.