Bug in Construct_initial_points if the non-bounding polyhedra are coplanar.
Before this patch, Construct_initial_points did not try to intersect with
the bounding polyhedron, but only with the other polyhedra. That can lead
to a crash if the given polyhedra are a bounding one, and a second
polyhedron that is coplanar. Then the initial points are *all* coplanar,
and then tr.dimension()==2 and an assertion is triggered.

Now Construct_initial_points uses construct_intersection_object(), and the
later tries to intersect with all polyhedra, including the bounding one.