*GMP_arithmetic_kernel now always define Integer and Rational even when Gmpfi is not available.
*Add default empty struct for Get_arithmetic_kernel. That way we can test if
a specialization is avaible. I used it for the default case of Root_of_traits.
When a number type not known by CGAL is used to instanciate Root_of_traits,
the field type used to instanciate Sqrt_extension is made using Quotient.

*Patch of Lazy_exact_nt constructor from another Lazy_exact_nt type.
In revision 61559, that constructor was retricted to the case the exact type
is implicitely constructed from the other exact type. Now the old behavior is
back but only if it is explicitely called (patch from Marc Glisse).

*Update To_double specialization of Sqrt_extension to use the same formula
than To_interval so that the double value is contained in the interval.

*update root_of_2 test-suite to use Sqrt_extension and that they behave
similarly