another cleanup: Another cleanup is to only #include what you need in this file.
For example fstream is not needed here.

   Removed all the unrelated header files. 

Avoid using statements.

 Removed "using namespace *** ".

Note that you make a copy of the  polyhedron,
that is you deform the copy not the original.
Not sure that this is your intention.

Yes this is my intention.  

Also intialize the data member in the initialization section
of the constructor

Do you mean add the initialization of roi,  hld and dsplc into constructor function?

I have no idea if #pragma once compiles with g++.

Replaced it by #ifndef....