Integration of a static filter for Do_intersect_3
So far, only operator()(BBox_3, Segment_3) has a static filter.

The result is gain by factor of 2 on a bench that does random segment
queries in an aabb tree of a real object (test/AABB_tree/data/finger.off).

The results on ten runs follow.

Without the static filter of Do_intersect_3:
| Epic kernel |10.21 | 
| Epic kernel |10.15 | 
| Epic kernel |10.46 | 
| Epic kernel |10.17 | 
| Epic kernel |10.12 | 
| Epic kernel |10.13 | 
| Epic kernel |10.16 | 
| Epic kernel |10.34 | 
| Epic kernel |10.2 | 
| Epic kernel |10.38 | 

With the static filter of Do_intersect_3()(BBox_3, Segment_3):
| Epic kernel |5.42 | 
| Epic kernel |5.4 | 
| Epic kernel |5.36 | 
| Epic kernel |5.41 | 
| Epic kernel |5.38 | 
| Epic kernel |5.41 | 
| Epic kernel |5.79 | 
| Epic kernel |5.53 | 
| Epic kernel |5.66 | 
| Epic kernel |5.45 |