Remove a problematic test that consumes too much CPU each night
Quote of a mail on mines:
"
We have problems with the testsuite of CGAL::Polytope_distance_d. The test 
test/Polytope_distance_d/test_Polytope_distance_d_d.cpp frequently enters an 
(seemly) infinite loop consuming CPU, in the testsuite, and is killed after 
our hour. On a given platform, the issue seems frequent but not fully 
reproducible.

Last night, that single test consumed 5 hours of CPU on my Linux test machine, 
and 2 hours on the Windows test machine, just to fail.
"