Recompile Boost libraries with g++-4.5
On that machine, I use g++-4.5, 4.6, and 4.7. If I compile the Boost
libraries with the oldest compiler, the binary compatibility is ensured,
because the soname of libstd++ is libstdc++.so.6 for all of those
compilers (but the binary compatibility is only ensured backward: if a
Boost library is compiled with g++-4.6, that does not work if it is after
linked with libstdc++ of g++-4.5).