New MPL type checking: is_streamable
is_streamable is a meta-function that checks if a type is streamable

is_streamable<T>::value is true iff the type T has stream operators <<
and >>. Otherwise it is false.


That commit adds a new header, used nowhere, and a new test. That cannot
break anything. I do it directly to 'next', with my own approval.