Fix mergeinfo properties
An attempt to simulate the reintegration of this branch into next gives
several errors.

One of them is:
svn: Reintegrate can only be used if revisions 63010 through 70034 were previously merged from svn+ssh://scm.gforge.inria.fr/svn/cgal/branches/next to the reintegrate source, but this is not the case:
  branches/features/Script-improve_create_cmake-eric/Maintenance/deb
    Missing ranges: /branches/CGAL-3.8-branch/Maintenance/deb:63010-63108

That branch is dead and does not concern anything in the current branch.

I fixed that with:
  svn merge -r 63009:63108 --record-only \^/branches/CGAL-3.8-branch/Maintenance/deb@63108

That only fixes the mergeinfo properties.