some problem with Coerction traits: Explicit casts required
Needs some further investigation! Background: the new CMake scripts for CGAL
are more strict what comes to includes of external libs (and Core is 
considered as an external lib now), thus the testsuite has shown errors 
related to the 'hacks' here. But as far as I know, the Coercion_traits 
should be able to solve these issues.