Increase the balls sizes when possible
This patch tries not to apply the general rule that the size of protecting
balls is a linear interpolation of the size of protecting balls at
corner. When the curve segment is long enough, pick a point at the middle
and choose a new size.

When insert_balls is called recursively, one needs to call its version with
all parameters. The reason is that d_sign is very important, if the curve
is a sub-curve of a cycle, otherwise that may loop!

CGAL_MESH_3_PROTECTION_NON_LINEAR is the default. 
One can still deactivate it with the macro
CGAL_MESH_3_NO_PROTECTION_NON_LINEAR.

(r64518, r64580, r64581, r64582 from Mesh_3-experimental-GF merged in one commit)