verbose=true validity test in sdg-info-set.cpp
I set the verbose=true variant of the test. However, the test
is of level zero in terms of deepness. This is in contrast with
the other programs in examples/Segment_Delaunay_graph_Linf_2/ and
probably should also be changed to level 1, which does not test
validity of the triangulation data structure. I have to talk to
Menelaos about this matter.