BUGFIX in offset computation
while computing the offset of a polygon, it might happen that a candidate polygonal
component is discarded (if it's not a valid polygon but a polyline for example).
In that case, some halfedges are marked as visited, but among them some could be
needed by another component. In this patch, we collect all halfedges marked as visited
and unmark them in case the polygon is incomplete.

The test added was not correctly working before the patch.