fix of subcurves assigned to top face
The correct thing is to re-assign subcurves starting on the left boundary
from the top face to another subcurve 'larger' in the status line when we
insert this larger curve. Only the topmost subcurve of the status line
after processing the left boundary should be assigned to the top face
We also added a new members (halfedge_indices_list) to helper classes and 
subcurve. It is open whether insertion/overlay visitor need the same fix