Proper namespace for ArrTraits and enable conceptification of namespaces.
conceptifying a namespace is the same as conceptify_nested_classes it
just has to be detected in another way if it is applicable.

For this purpose the macro \cgalconceptnamespace has been added. It
can be used the same way as \cgalconcept to mark the documentation of
a namespace for conceptification. For clarity some documentation of
the namespace should be provided.

It is not possible to mix concepts and classes in the same namespace.