Vertex conflict pqt + Construction pcl
When t is segment, p is endpoint of segment
and q is not end point of segment, then the vertex at infinity
between p and q should be in conflict, we have updated this
in this commit.
Also we have done a little change in Construction of pcl
between a endpoint of segment and segment itself.
we have omitted the portion that if q is point we have to change
the direction, this is because the logic is written wrt pnt and seg,
so it became independent of p and q.