Removed a rare data race condition.
All the initial tasks (i.e. the tasks pushed into the priority queue by initialize_cells_priority_queue) need to be "spawned" before we really start to run them on the CPU.
For now, I use an atomic bool to force the tasks to wait, but I will replace this by a better solution very soon.