Fix the memory leak in minkowski_sum_3
Gaussian_map allocates a Sphere_map on every construction and passes
it off to its base class SM_decorator. The decorator has a shallow
copy-constructor and doesn't clean up its resource.

The best fix would be to use a shared_ptr for psm_ in SM_decorator,
but that possibly breaks users and classes that actually account for
the missing memory management.

This is the worst of all possible fixes. We just clean up all
pointers. The code now checks fine with gperftools and
HEAPCHECK=strict