bugfix of set_coef introduced by c84cc28d1c5550876d2d648f905716ffa69335ef
The problem is that building the matrix from a set of triplets sums
the value provided in case several values at the same position are
provided. In order to overwrite a value, we have no other choice
than to build the matrix and set the value (with current Eigen API).
The following fix is as efficient if the matrix is assembled
in one pass. Using the boolean member of set_coef is very important now
as it can imply a premature building of the eigen matrix.