ipescript ipe2cin converts ipe to cin format
The script ipe2cin converts all segments and points
in the first page of the ipe file given as the command
line argument to the SDG Linf cin format for segments
and points. The result is printed to the standard output.
Use refirection to save to a .cin file.

The file ipe2cin.lua has to be put in the appropriate
place, e.g., ~/.ipe/scripts/ and here is an example use:

ipescript ipe2cin ipe1.ipe > ipe1.cin

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>