Remove the variable Boost_DIR from the CMake cache
The CMake module FindBoost.cmake searchs for Boost first in the "Config
mode" of 'find_package'. That allows to find installation of boost-cmake
(that is Boost configured using CMake). That configuration of Boost is
rather rare. But the search for it leaves a variable Boost_DIR=NOTFOUND in
the cache, even if Boost libraries are found using the regular way.

That patch removes from the cache the variable Boost_DIR, when Boost is
already found, to avoid more confusion for users.