Update the configuration of picasso
The configuration was already updated for a while, but not commited.

Today I have added a new platform that will use LEDA instead of GMP. And I
have disable x64_Cygwin-Windows8_MSVC2012-Debug-32bits

That commit also adds a shell script to update the local copy from the
files on picasso.