template parameter of SDG graph in SDG hierarchy
Use an additional optional parameter (SDGLx) in the segment
Delaunay hierarchy template to be able to change the segment
Delaunay graph class that is used per level of the hierarchy.
Since this is the last optional parameter, there is no change
for users of the L2 segment Delaunau graph and hierarchy.

This parameter is necessary to support the Linf segment Delaunay
graph and hierarchy. Again, it is intended that the user of the
Linf classes does not have to be aware of this parameter.

Some additional classes are declared as friends of
Segment_Delaunay_graph_2.