endpoint of segment code completion
case of predicate pqrt of type PPSP:
if new point t is on the same side of line pq
as the other endpoint of r, then NEGATIVE, else POSITIVE.

This fixes inputs of the following form (zppsp1.cin):
s -100 50 0 0
p 100 0
p 0 100

With the fix, point p 0 0 has as neighbors both
p 100 0 and p 0 100. Before, p 0 100 was not a neighbor.

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>