Remove a CMake warning when building from the release tarball
== Setting paths ==
-- Build CGAL from release in directory CGAL-4.3-beta1
CMake Warning (dev) at CMakeLists.txt:99 (set):
  Cannot set "CGAL_CONFIGURED_PACKAGES_NAMES": current scope has no parent.
This warning is for project developers.  Use -Wno-dev to suppress it.

The reason is that in the release there is not parent directory. As that
cmake instruction is need only if the Doxygen documentation is there, I
have surrounded the corresponding line of code by:
  if(IS_DIRECTORY "${CMAKE_SOURCE_DIR}/Documentation")
  endif()