Trivial bug-fix: use FT(n) instead of (FT)n.0
In our concepts, (FT)0.0 and (FT)1.0 (conversion from doubles) are not
always allowed, but the construction of an FT from an integer is in the
concept FieldNumberType (because it refines IntegralDomainWithoutDivision,
that refines FromIntConstructible).

The testsuite of PCA passes on my machine, and the bug "fix" is trivial.
Approved for CGAL-4.3 by the Release Manager.