trivial bugfix preventing segfault in case of filter failure
The construction of the line l can fail while the size of the list
has been increased. This would lead to trying to remove an element
never added. The fix simply delay the size increment after the line
construction.

the construction of the line can fail and the new point no