Add destructor and delete allocated memory
In the only use case in the code base SAVED_OBJECT is a
pointer. Technically the deallocation of the objects should happen
inside the snap rounding algorithm itself, but this would some
modifications to the algorithm necessary. For simplicity
Multiple_kd_tree now takes ownership of the SAVED_OBJECTs and
deallocates them.