fix memory leak in some Gmpfi operators
The leak was caused by the fact that some operators created an mpfi_t to
store the result and returned a Gmpfi constructed from that mpfi_t. The
problem was that it was impossible to free the memory used by that
number. The fix consisted in storing the result directly in a Gmpfi,
instead of a mpfi_t (this also results in a cleaner code).